Continuing troubles with JDK16 asserting itself
Not sure why a couple of hours worth of experiments still fails to get rid of the missing tools-1.4.2.jar issue, after even putting it in the m2 cache myself, and rebooting after making sure java home is set to 1.8 vs. 16. But it keeps saying it isn't in the JDK 16 dir, which we already know. So this is a placeholder checkin as I'm about to delete that new dep from the POM as it wasn't there before and this might be a bogus error now that I have java home set correctly again.