Prepared the DXF converter code for extraction to a new library
The DXF converter code that takes a loaded DXF file parsed into JavaFX graphics objects and converts it to JavaFX that is ready to load in a GUI or Scene Graph, does not use jfxconverter, so it will be split into its own FxDxfConverter library tomorrow, so it can be built and used. It will stay separate even once jfxconverter is in some form that it can be built into GitHub Maven-based libraries.