Apparent success in getting jfxconverter dependencies to resolve finally
I will update the README.TXT to instruct on how to mimic Maven Central for the three external JAR files that are required from JFXConverter. A bit of a hack, but it was easier than trying to integrate with a recent third-party hack that tries to make the old Ant-built SourceForge project work with Maven. I will be in contact with that developer to see if we can combine forces.